+++ src/examples/index.ts
@@ -1,3 +1,2 @@
 export * from "./basic.js";
-export * from "./enhanced.js";
 export * from "./auto.js";