+++ src/index.ts
@@ -2,7 +2,6 @@ import { ElfaSDK } from "./client/ElfaSDK.js";
 
 export { ElfaSDK } from "./client/ElfaSDK.js";
 export { ElfaV2Client } from "./client/ElfaV2Client.js";
-export { TwitterClient } from "./client/TwitterClient.js";
 export { AutoClient } from "./client/AutoClient.js";
 export { TradeClient } from "./client/TradeClient.js";
 
@@ -11,12 +10,7 @@ export * from "./types/index.js";
 export * from "./utils/errors.js";
 export { PaginationHelper } from "./utils/pagination.js";
 
-export type {
-  SDKOptions,
-  RequestOptions,
-  EnhancementOptions,
-  EnhancedResponse,
-} from "./types/enhanced.js";
+export type { SDKOptions } from "./types/options.js";
 
 export type {
   TrendingTokensParams,
@@ -27,10 +21,5 @@ export type {
   TopMentionsV2Params,
   TrendingNarrativesParams,
 } from "./types/elfa.js";
-
-export type {
-  TwitterClientOptions,
-  TwitterBatchRequest,
-} from "./types/twitter.js";
 
 export default ElfaSDK;