+++ src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from "./errors.js";
 export * from "./http.js";
-export * from "./enhancer.js";
 export * from "./pagination.js";